Add files via upload
Mediapipe Posenet Library utilised with additional code assigning the co-ordinates of each point of interest (pose landmark) to a variable, displaying these variables to the user. Additionally the code added ascertains the users position in the frame, with placeholder commands to be implemented in ROS to turn towards the user based on their position, moving forwards and stopping if they get within a certain distance of the user.